+++ CHANGELOG.md
@@ -49,6 +49,7 @@
   - `TestEmpty()` now supports tracing
   - `TestFilepathExists()` now supports tracing
   - `TestNotEmpty()` now supports tracing
+  - `ToStderr()` now supports tracing
   - `Tr()` now supports tracing
   - `TrimSuffix()` now supports tracing
   - `TrimWhitespace()` now supports tracing